remove snapshot

+++ catalog/models.py
@@ -1,14 +1,5 @@
 from django.db import models
 from django.dispatch.dispatcher import receiver
-import os
-
-class Snapshot(models.Model):
-    file = models.ImageField(upload_to='snapshots', default='default.jpg')
-
-@receiver(models.signals.pre_delete, sender=Snapshot)
-def auto_delete_snapshot(sender, instance, **kwargs):
-    if instance.file:
-        instance.file.delete()
 
 class VideoFile(models.Model):
     file = models.FileField(upload_to='videos', default='default.mp4')
@@ -23,7 +14,6 @@ def get_default_videofile():
 
 class Video(models.Model):
     name = models.CharField(max_length=255)
-    snapshot = models.OneToOneField(Snapshot, on_delete=models.CASCADE, null=True)
     videofile = models.OneToOneField(VideoFile, on_delete=models.CASCADE, null=False, default=get_default_videofile)
 
     def __str__(self):
@@ -31,7 +21,5 @@ class Video(models.Model):
 
 @receiver(models.signals.pre_delete, sender=Video)
 def auto_delete_video(sender, instance, **kwargs):
-    if instance.videofile.file:
+    if instance.videofile and instance.videofile.file:
         instance.videofile.file.delete()
-    if instance.snapshot.file:
-        instance.snapshot.file.delete()
